==> R_scripts/Bird-location/Step3_ModelEvaluation.Rmd <==
***[Bird Location BRT Model Evaluation]***
For Bird-location Oregon 2020 data
Goal: Evaluate species distribution models (CV statistics) for all species of birds from Oregon trained BRT

Load packages
```{r}
library(pROC)
library(foreach)
library(doParallel)
library(reticulate)
library(rgee)
library(ggplot2)
library(mice)
library(gbm)
library(terra)
library(raster)
library(dismo)
library(gbm)
library(tidyverse)
library(readr)
library(Metrics)
```

Load Test Data for bird-location
```{r}
# Define the base directory
testdata_dir <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/TestData/SpeciesCentered/"

# Define file names
file_names <- c("SpeciesCentered_FixedRadius.csv", "SpeciesCentered_noRadius.csv", "SpeciesCentered_ESW.csv")

# Load and assign datasets to variables
fixed.test <- read.csv(paste0(testdata_dir, file_names[1]))
noRadius.test <- read.csv(paste0(testdata_dir, file_names[2]))
esw.test <- read.csv(paste0(testdata_dir, file_names[3]))

head(fixed.test)
```

Load the Trained models for Bird-location strategy
```{r}
individual_species_oregon <- split(fixed.test, fixed.test$Common_Name)

model.dir.noradius <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/TrainedModel/SpeciesCentered/SpeciesCentered_NoRadius"
noradius.models <- list.files(model.dir.noradius, pattern = "\\.rds$", full.names = TRUE, recursive = TRUE)

model.dir.fixedradius <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/TrainedModel/SpeciesCentered/SpeciesCentered_FixedRadius"
fixedradius.models <- list.files(model.dir.fixedradius, pattern = "\\.rds$", full.names = TRUE, recursive = TRUE)

model.dir.esw <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/TrainedModel/SpeciesCentered/SpeciesCentered_ESW"
esw.models <- list.files(model.dir.esw, pattern = "\\.rds$", full.names = TRUE, recursive = TRUE)

```



***===Model Evaluation (CV statistic, Kendall)====***
We will use deviance to test difference of likelihood between the fitted model (BRTs) and the saturated model
- First Assign NA to NULL BRT models for each species & predictor extraction method
Load all models for a given species and calculate deviance values
```{r}
# Create empty dataframe to store evaluation results
result <- data.frame(Species = character(),
                     Extraction_Approach = character(),
                     Metric = character(),
                     Pixel_radius = numeric(),
                     Fixed_Radius = numeric(),
                     Effective_Radius = numeric())

load_models_for_species <- function(species_name) {
    
    # Initialize variables to store Kendall's rank coefficient values
    kendall_noRadius <- NA
    kendall_fixed <- NA
    kendall_esw <- NA
    
    # Initialize variables to store CV stats
    cv_meandeviance_noRadius <- NA
    cv_meandeviance_fixed <- NA
    cv_meandeviance_esw <- NA
    
    cv_SEdeviance_noRadius <- NA
    cv_SEdeviance_fixed <- NA
    cv_SEdeviance_esw <- NA

    
    #=== For pixel radius ===#
    noradius.perspecies <- grep(species_name, noradius.models, value = TRUE)
    
    if (length(noradius.perspecies) > 0) {
        noradius.perspecies.model <- readRDS(noradius.perspecies)
        
        if (!is.null(noradius.perspecies.model$gbm.call$best.trees)) {
            pred.noRadius <- predict.gbm(object = noradius.perspecies.model, 
                                         newdata = noRadius.test[noRadius.test$Common_Name == species_name, ],
                                         n.trees = noradius.perspecies.model$gbm.call$best.trees,
                                         type = "response")
            
            # Calculate Kendall's correlation
            kendall_noRadius <- cor(x = noRadius.test[noRadius.test$Common_Name == species_name, ]$Occur, 
                                    y = pred.noRadius, method = "kendall")
            
            # Get CV statistics from trained model itself
            cv_meandeviance_noRadius <- noradius.perspecies.model$cv.statistics$deviance.mean
            cv_SEdeviance_noRadius <- noradius.perspecies.model$cv.statistics$deviance.se
            
        }
    }

    #=== For fixed radius ===#
    fixedradius.perspecies <- grep(species_name, fixedradius.models, value = TRUE)
    
    if (length(fixedradius.perspecies) > 0) {
        fixedradius.perspecies.model <- readRDS(fixedradius.perspecies)
        
        if (!is.null(fixedradius.perspecies.model$gbm.call$best.trees)) {
            pred.fixed <- predict.gbm(object = fixedradius.perspecies.model, 
                                      newdata = fixed.test[fixed.test$Common_Name == species_name, ],
                                      n.trees = fixedradius.perspecies.model$gbm.call$best.trees,
                                      type = "response")
            
            # Calculate Kendall's correlation
            kendall_fixed <- cor(x = fixed.test[fixed.test$Common_Name == species_name, ]$Occur, 
                                 y = pred.fixed, method = "kendall")
            
            # Get CV statistics from trained model itself
            cv_meandeviance_fixed <- fixedradius.perspecies.model$cv.statistics$deviance.mean
            cv_SEdeviance_fixed <- fixedradius.perspecies.model$cv.statistics$deviance.se

        }
    }

    #=== For Effective radius ===#
    esw.perspecies <- grep(species_name, esw.models, value = TRUE)
    
    if (length(esw.perspecies) > 0) {
        esw.perspecies.model <- readRDS(esw.perspecies)
        
        if (!is.null(esw.perspecies.model$gbm.call$best.trees)) {
            pred.ESW <- predict.gbm(object = esw.perspecies.model, 
                                    newdata = esw.test[esw.test$Common_Name == species_name, ],
                                    n.trees = esw.perspecies.model$gbm.call$best.trees,
                                    type = "response")
  
            # Calculate Kendall's correlation
            kendall_esw <- cor(x = esw.test[esw.test$Common_Name == species_name, ]$Occur, 
                                 y = pred.ESW, method = "kendall")

            # Get CV statistics from trained model itself
            cv_meandeviance_esw <- esw.perspecies.model$cv.statistics$deviance.mean
            cv_SEdeviance_esw <- esw.perspecies.model$cv.statistics$deviance.se

        }
    }

    # Store results in a dataframe

      # Kendall rank correlation test
    per_bird.result.kendall <- data.frame(Species = species_name,
                                      Extraction_Approach = "Bird-location Approach",
                                      Metric = "Kendall",
                                      Pixel_radius = kendall_noRadius,
                                      Fixed_Radius = kendall_fixed,
                                      Effective_Radius = kendall_esw)
      # CV statistics
        # Mean deviance
    per_bird.result.CVdeviance <- data.frame(Species = species_name,
                                             Extraction_Approach = "Bird-location Approach",
                                             Metric = "CV_MeanDeviance",
                                             Pixel_radius = cv_meandeviance_noRadius,
                                             Fixed_Radius = cv_meandeviance_fixed,
                                             Effective_Radius = cv_meandeviance_esw)
        # SE deviance
    per_bird.result.CV.SEdeviance <- data.frame(Species = species_name,
                                                Extraction_Approach = "Bird-location Approach",
                                                Metric = "CV_SE_Deviance",
                                                Pixel_radius = cv_SEdeviance_noRadius,
                                                Fixed_Radius = cv_SEdeviance_fixed,
                                                Effective_Radius = cv_SEdeviance_esw)
    
    # Concatenate per bird results in one data frame
    per_bird.result <- rbind(per_bird.result.kendall,
                             per_bird.result.CVdeviance, per_bird.result.CV.SEdeviance)
    
    return(per_bird.result)
}


# Loop over each bird species name and load models
for (bird in names(individual_species_oregon)) {
  print(bird)
  per_bird.result <- load_models_for_species(bird)
  result <- rbind(result, per_bird.result)
}

```

Export as CSV
```{r}
write.table(result, file = "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/Evaluation/Oregon2020/evalMetric-BirdLocationOR2020.csv", sep = ",", row.names = FALSE, fileEncoding = "UTF-8")

```
